+++ Artwalk_Android/domain/src/main/java/com/a401/domain/model/RouteForDraw.kt
@@ -2,12 +2,12 @@ package com.a401.domain.model
 
 data class RouteForDraw(
     val duration: Int,
-    val distance: Float,
+    val distance: Double,
     val geometry: String,
 ) {
     constructor() : this(
         0,
-        0.0f,
+        0.0,
         "",
     )
 }
